1. Format of table is wrong in code
The format of table is wrong when displayed in code.

Table format has been readjusted for easier reading and editing.

2. Missing extensions for Mark

Extensions for the Mark use case is not complete.

Added extensions for Mark use case if user tries to find a reservation
that does not exist or tries to mark and already marked reservation

3. Missing Unmark use case

Unmark use case not added yet.

Added unmark use case together with relevant extensions.